+++ src/main/java/it/impresaconsulting/Gestic/daos/ScadenzaDao.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ScadenzaDao extends JpaRepository<Scadenza, Integer>{
+public interface ScadenzaDao extends JpaRepository<Scadenza, Long>{
     List<Scadenza> findByRegistratoDa(String registratoDa);
 }
